refactor: share map tile-marker sort and split fleet/DTO hosts (#4654)
Slice A gives army, fleet, and civilian markers one y/x/tileKey sort, one human-id helper, ProvinceUnitPresenceView.copyWith, and a repo lint so those sorts cannot be re-inlined under view/.
