+++ SPEC/program/repo-lint.md
@@ -194,6 +194,7 @@
 | `tool/check_map_test_minimal_game_shared.dart` | Forbid inline `\bGame\(` in the scoped colour/capital/region-input/format/port suites (`faction_ownership_color_test.dart`, `tile_map_capital_markers_test.dart`, `region_map_view_inputs_test.dart`, `map_format_util_test.dart`, `port_icon_placement_test.dart`); those suites must use `minimalGame` from `test/support/init_game_map_view_fixtures.dart`. Full-line comments are stripped; `minimalGame(` is not matched (Refs #4022); rule `repo.map_test_minimal_game_shared` |
 | `tool/check_map_test_file_size.dart` | Physical-line ceiling **250** for every `packages/colonizethis_map/test/**` Dart file unless manifest-baseline grandfathered (Refs #4112, #4561); rule `repo.map_test_file_size` |
 | `tool/check_map_render_legend_layout_dedup.dart` | Forbid inline `legendPadding * 2 + legendLines * legendLineHeight` outside `tile_map_visualization_legend_layout.dart` (canonical `legendHeightForLineCount` owner); full-line comments stripped (Refs #4112); rule `repo.map_render_legend_layout_dedup` |
+| `tool/check_map_tile_marker_sort_sot.dart` | Forbid inlined y then x then `tileKey` sorts under `packages/colonizethis_map/lib/src/view/**` outside `init_game_map_view_tile_marker_sort.dart` (Refs #4654); rule `repo.map_tile_marker_sort_sot` |
 | `tool/check_combat_test_mirrors_lib.dart` | Keep the `packages/colonizethis_combat/test/**` tree mirrored to `lib/` per the testing rule (`test/` mirrors `lib/`): any Dart source that sits **directly** at the `test/` root (its repo-relative path after `packages/colonizethis_combat/test/` contains no further `/`) is a violation — both `*_test.dart` files and co-located support helpers must live in a subdirectory such as `test/combat/` that mirrors `lib/src/combat/`. Files already nested under a subdirectory (e.g. `test/combat/combat_rng_test.dart`) are compliant; generated suffixes and fixture trees are skipped via the shared scan contract. Complements the testing-rule `test/`-mirrors-`lib/` policy (Refs #3716); rule `repo.combat_test_mirrors_lib` |
 | `tool/check_combat_test_no_local_support.dart` | Forbid package-local `*_test_support.dart` files under `packages/colonizethis_combat/test/**`: canonical home is `colonizethis_combat_test_support` (Refs #3865); rule `repo.combat_test_no_local_support` |
 | `tool/check_combat_test_core_fixtures_shared.dart` | Forbid inline `Game(` construction under `packages/colonizethis_combat/test/**`; use shared builders from `colonizethis_combat_test_support` (Refs #3865); rule `repo.combat_test_core_fixtures_shared` |
