+++ packages/colonizethis_map/lib/src/view/init_game_map_view_fleet_markers.dart
@@ -1,154 +1,19 @@
 /// Fleet tile markers and in-port ship-count overlays for map view data.
-/// SPEC/program/map-visualization.md § Map view model for tools. Refs #4022.
+/// SPEC/program/map-visualization.md § Map view model for tools. Refs #4022, #4654.
 library;
 
 import 'package:colonizethis_data/colonizethis_data.dart';
-import 'package:colonizethis_map/package_logger.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 
-import '../map_pipe_string_util.dart';
-import '../port_icon_placement.dart';
-import '../sea_zone_centroid_tile.dart';
-import '../tile_key_util.dart';
 import 'init_game_map_view_data.dart';
+import 'init_game_map_view_fleet_markers_inclusion.dart';
+import 'init_game_map_view_fleet_markers_location.dart';
+import 'init_game_map_view_human_player_ids.dart';
+import 'init_game_map_view_tile_marker_sort.dart';
 
 /// Fleet stacking / in-port placement markers for a single region's view data.
 class InitGameMapViewFleetMarkers {
   const InitGameMapViewFleetMarkers._();
-
-  static final _log = packageLogger();
-
-  static String _homeFleetIdForMapMarker(String playerId) => 'fleet_$playerId';
-
-  static bool _fleetAtHumanCapital(Game game, String playerId, Fleet fleet) {
-    if (!fleet.isInPort || fleet.inPortAtProvinceId == null) {
-      return false;
-    }
-    final player = game.players.firstWhere(
-      (p) => p.id == playerId,
-      orElse: () => game.players.first,
-    );
-    if (!player.isHuman) {
-      return false;
-    }
-    final cap = player.capitalTile;
-    if (cap == null) {
-      return false;
-    }
-    final parsed = tryParseMapTileKey(cap.toTileKey());
-    if (parsed == null) {
-      return false;
-    }
-    final capReg = parsed.regionId;
-    final capProvLocal = parsed.localId;
-    if (fleet.regionId != capReg) {
-      return false;
-    }
-    final port = fleet.inPortAtProvinceId!;
-    return port == capProvLocal || port == '$capReg|$capProvLocal';
-  }
-
-  static bool _includeFleetForTileMarker(
-    Game game,
-    Fleet f,
-    String regionId,
-    Set<String> humanIds,
-  ) {
-    if (!humanIds.contains(f.ownerId) || f.regionId != regionId) {
-      return false;
-    }
-    if (f.shipTypeIds.isNotEmpty) {
-      return true;
-    }
-    return f.id == _homeFleetIdForMapMarker(f.ownerId) &&
-        _fleetAtHumanCapital(game, f.ownerId, f);
-  }
-
-  static String? _inPortFleetMarkerTileKey({
-    required Game game,
-    required String regionId,
-    required Province province,
-    required TileMapResult tileMap,
-    required Set<String> seaZoneIds,
-  }) {
-    final localProvinceId = ProvinceId.localIdFrom(province.id);
-    final tileKey = harborDrawableSeaTileKeyForPortProvince(
-      game: game,
-      regionId: regionId,
-      localProvinceId: localProvinceId,
-      tileMap: tileMap,
-      seaZoneIds: seaZoneIds,
-      contextLabel: 'fleet marker region=$regionId province=$localProvinceId',
-    );
-    if (tileKey == null) {
-      _log.w(
-        'map: in-port fleet marker skipped: no portsByProvinceSeaboard entry '
-        'for region=$regionId province=$localProvinceId',
-      );
-    }
-    return tileKey;
-  }
-
-  static (int?, int?) _xyFromMapTileKey(String tileKey) {
-    final parsed = tryParseMapTileKeySuffixXY(tileKey);
-    if (parsed == null) {
-      return (null, null);
-    }
-    return (parsed.x, parsed.y);
-  }
-
-  static String? _fleetMarkerTileKeyForLocationScope({
-    required String scopeKey,
-    required String regionId,
-    required Game game,
-    required TileMapResult tileMap,
-    required Set<String> seaZoneIds,
-    required Map<String, Province> provinceMap,
-  }) {
-    if (scopeKey.startsWith('sea:')) {
-      final zoneKey = scopeKey.substring(4);
-      final local = mapPipeLastSegmentOrWhole(zoneKey);
-      return seaZoneCentroidTileKey(
-        tileMap: tileMap,
-        regionId: regionId,
-        localSeaZoneId: local,
-        seaZoneNodeIds: seaZoneIds,
-      );
-    }
-    if (!scopeKey.startsWith('port:')) return null;
-    final fullProv = scopeKey.substring(5);
-    final province = provinceMap[fullProv];
-    if (province == null) return null;
-    return _inPortFleetMarkerTileKey(
-      game: game,
-      regionId: regionId,
-      province: province,
-      tileMap: tileMap,
-      seaZoneIds: seaZoneIds,
-    );
-  }
-
-  static void _addFleetToLocationBuckets({
-    required Fleet f,
-    required String regionId,
-    required Map<String, Province> provinceMap,
-    required Map<String, List<Fleet>> byLocation,
-  }) {
-    if (f.isAtSea && f.seaZoneId != null) {
-      final z = f.seaZoneId!;
-      final zoneKey = z.contains('|') ? z : '$regionId|$z';
-      byLocation.putIfAbsent('sea:$zoneKey', () => []).add(f);
-      return;
-    }
-    if (f.inPortAtProvinceId == null) return;
-    final province =
-        provinceMap['$regionId|${f.inPortAtProvinceId}'] ??
-        provinceMap[f.inPortAtProvinceId!];
-    if (province == null) return;
-    byLocation
-        .putIfAbsent('port:${province.regionId}|${province.id}', () => [])
-        .add(f);
-  }
 
   static List<FleetTileMarkerView> buildFleetTileMarkersForRegion({
     required Game game,
@@ -157,10 +22,7 @@ class InitGameMapViewFleetMarkers {
     required TileMapResult tileMap,
     required Set<String> seaZoneIds,
   }) {
-    final humanIds = game.players
-        .where((p) => p.isHuman)
-        .map((p) => p.id)
-        .toSet();
+    final humanIds = humanPlayerIds(game);
     if (humanIds.isEmpty) {
       return const [];
     }
@@ -172,11 +34,11 @@ class InitGameMapViewFleetMarkers {
     final byLocation = <String, List<Fleet>>{};
 
     for (final f in game.worldState.fleets) {
-      if (!_includeFleetForTileMarker(game, f, regionId, humanIds)) {
+      if (!includeFleetForTileMarker(game, f, regionId, humanIds)) {
         continue;
       }
-      _addFleetToLocationBuckets(
-        f: f,
+      addFleetToLocationBuckets(
+        fleet: f,
         regionId: regionId,
         provinceMap: provinceMap,
         byLocation: byLocation,
@@ -189,7 +51,7 @@ class InitGameMapViewFleetMarkers {
       final fleets = entry.value.toList()..sort((a, b) => a.id.compareTo(b.id));
       final fleetIds = fleets.map((fl) => fl.id).toList();
 
-      final tileKey = _fleetMarkerTileKeyForLocationScope(
+      final tileKey = fleetMarkerTileKeyForLocationScope(
         scopeKey: scopeKey,
         regionId: regionId,
         game: game,
@@ -200,7 +62,7 @@ class InitGameMapViewFleetMarkers {
       if (tileKey == null) {
         continue;
       }
-      final xy = _xyFromMapTileKey(tileKey);
+      final xy = xyFromMapTileKey(tileKey);
       final x = xy.$1;
       final y = xy.$2;
       if (x == null || y == null) {
@@ -217,17 +79,12 @@ class InitGameMapViewFleetMarkers {
         ),
       );
     }
-    markers.sort((a, b) {
-      final yc = a.y.compareTo(b.y);
-      if (yc != 0) {
-        return yc;
-      }
-      final xc = a.x.compareTo(b.x);
-      if (xc != 0) {
-        return xc;
-      }
-      return a.tileKey.compareTo(b.tileKey);
-    });
+    sortTileAnchoredMarkers(
+      markers,
+      yOf: (m) => m.y,
+      xOf: (m) => m.x,
+      tileKeyOf: (m) => m.tileKey,
+    );
     return markers;
   }
 
@@ -248,11 +105,8 @@ class InitGameMapViewFleetMarkers {
       if (current == null) {
         continue;
       }
-      provincePresenceById[provinceId] = ProvinceUnitPresenceView(
-        civilianCount: current.civilianCount,
-        regimentCount: current.regimentCount,
+      provincePresenceById[provinceId] = current.copyWith(
         shipCount: current.shipCount + fleet.ships.length,
-        intelVisible: current.intelVisible,
       );
     }
   }
