+++ tool/ct_repo_lint_manifest.yaml
@@ -2214,6 +2214,13 @@ rules:
     runner: dart
     script: tool/check_map_render_legend_layout_dedup.dart
 
+  - rule_id: repo.map_tile_marker_sort_sot
+    group: architecture
+    title: colonizethis_map view markers must use shared y/x/tileKey sort helper (Refs #4654)
+    spec: SPEC/program/repo-lint.md
+    runner: dart
+    script: tool/check_map_tile_marker_sort_sot.dart
+
   - rule_id: repo.combat_test_mirrors_lib
     group: architecture
     title: colonizethis_combat tests must mirror lib/ (no loose Dart sources at the test/ root) (Refs #3716)
